rmdir only if extracted

+++ getXS.py
@@ -10,9 +10,12 @@ if __name__ == '__main__':
     parser.add_argument('-o', dest='outfile', help='Output path', required=True)
     parser.add_argument('--ext', dest='ext', help='Archives extension', required=False, default='tar.gz')
     args = parser.parse_args()
+
+    rmdir = False
     
     tars = glob(os.path.abspath(args.base) + '/*.' + args.ext)
     for tar in tars:
+
         cfgpath = os.path.basename(tar).split('.')[0] + '/read_03_input.cfg'
         cfgdir = os.path.dirname(os.path.abspath(cfgpath))
         if '_results' not in cfgdir:
@@ -27,13 +30,18 @@ if __name__ == '__main__':
                     continue
             else:
                 print('[INFO] working on ' + tar)
+                rmdir = True
                 os.system('tar --extract --file={0} {1}'.format(tar, cfgpath))
                 print('       read_03_input.cfg extracted')
+
         with open(cfgpath, 'r') as cp:
             cfg = configparser.ConfigParser()
             cfg.readfp(cp)
             sample_name = [str(c) for c in cfg.sections() if str(c).startswith('Z')][0]
             sample_XS = str(cfg.get(sample_name, 'XS')) # pb
-        with open(args.outfile, 'w+') as o:
+            
+        with open(args.outfile, 'a') as o:
             o.write(sample_name + '\t' + sample_XS + '\n')
-        os.system('rm -rf ' + cfgdir)
+
+        if rmdir:    
+            os.system('rm -rf ' + cfgdir)
